i18n_seed_pipeline/requirements.txt
requests>=2.32.3
sqlparse>=0.5.1
tqdm>=4.66.5
pyyaml>=6.0.2
